requirements.txt
psutil
fastapi
uvicorn[standard]
psycopg2-binary
